Removing theory units option to avoid confusion. Use direct scaling in theory expression. Also TermInfo option is introduced in the latter.
git-svn-id: https://svnsrv.desy.de/desy/h1fitter/trunk@1873 b2a8d302-ca0a-427a-b7b9-081adbce06bc

+++ include/CommonGrid.h
@@ -48,7 +48,7 @@ class CommonGrid {
 
   //! Selects if we have a proton-antiproton collision
   void SetCollisions(int ppbar) {_ppbar = (ppbar == 1);};
-  void SetDynamicScale(float dynscale) {_dynamicscale = dynscale;};
+  void SetDynamicScale(double dynscale) {_dynamicscale = dynscale;};
   //! Check that the data and grid bins are consistent
   int checkBins(vector<int> &bin_flags, vector<vector<double> > &data_bins);
 
@@ -94,7 +94,7 @@ class CommonGrid {
   bool _ppbar;
 
   /// bin-by-bin dynamic scale
-  float _dynamicscale;
+  double _dynamicscale;
 };
 
 #endif
